requirements.txt
google-genai
python-dotenv
pymupdf
Pillow
jinja2
streamlit